Move thrift-specific classes into m3 package
